% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.inflmixNPMLE}
\alias{print.inflmixNPMLE}
\title{Summarize the results of several estimated OIPPMMs.}
\usage{
\method{print}{inflmixNPMLE}(x, ...)
}
\arguments{
\item{x}{An object of class "\code{inflmixNPMLE}", a result of a call to
\code{\link{inflmix}} with \code{l=NULL} and \code{p=NULL}, and with
\code{K=NULL}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Custom print method for objects of class \code{inflmixNPMLE}.
}
\description{
Summarize the results of several estimated OIPPMMs.
}
\examples{
inflmix(1:20)
}
\seealso{
\code{\link{inflmix}}
}
