% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinflmix.R
\name{rinflmix}
\alias{rinflmix}
\title{Generate variates from the OIPPMM.}
\usage{
rinflmix(n, l, p)
}
\arguments{
\item{n}{the desired sample size}

\item{l}{lambda, a vector of positive Poisson mixture components}

\item{p}{pi, a vector of weights for the mixture components}
}
\value{
returns a list containing \code{y} (the generated variates) and \code{n0}
(the number of 0s from the Poisson process that were discarded)
}
\description{
Generate variates from the OIPPMM.
}
\examples{
y <- rinflmix(n=50, l=c(1,4), p=c(.4,.4))$y
sumOIPPMM(inflmix(y))
}
\seealso{
\code{\link{inflmix}}, \code{\link{sumOIPPMM}}
}
