% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinflmixN.R
\name{rinflmixN}
\alias{rinflmixN}
\title{Generate \code{n} variates from the OIPPMM, beginning with unknown population size, \code{N}.}
\usage{
rinflmixN(N, l, p)
}
\arguments{
\item{N}{the unknown population size}

\item{l}{lambda, a vector of positive Poisson mixture components}

\item{p}{pi, a vector of weights for the mixture components}
}
\value{
a vector of generated OIPPMM variates
}
\description{
Generate \code{n} variates from the OIPPMM, beginning with unknown population size, \code{N}.
}
\examples{
y <- rinflmixN(N=500, l=c(1,4), p=c(.4,.4))
inflmix(y)
}
\seealso{
\code{\link{inflmix}}, \code{\link{rinflmix}}
}
