% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumOIPPMM.R
\name{sumOIPPMM}
\alias{sumOIPPMM}
\title{Summarize the results of several estimated OIPPMMs.}
\usage{
sumOIPPMM(zz)
}
\arguments{
\item{zz}{An object of class "\code{inflmixNPMLE}", a result of a call to
\code{\link{inflmix}} with \code{l=NULL} and \code{p=NULL}.}
}
\value{
Prints a summary matrix of strings.
}
\description{
Summarize the results of several estimated OIPPMMs.
}
\examples{
require(maxLik)
# Search for the non-parametric MLE
a <- inflmix(1:20)
# Summarize information about the various models fit in the search for the NPMLE
sumOIPPMM(a)
}
\seealso{
\code{\link{inflmix}}
}
